reorganize evm interfaces/structs
